+++ Sphinx/ChangeLog.rst
@@ -1,6 +1,8 @@
 ChangeLog
 =========
-* December 31, 2014
+* January 1st, 2015
+
+  - Added ``--fps n`` option.
 
   - Added Donate button in ``_draw_about()`` panel.